pyproject.toml
[project]
name = "vitaura"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "gradio-client>=1.13.3",
    "httpx>=0.28.1",
    "langchain>=1.0.2",
    "langchain-core>=1.0.1",
    "langchain-google-genai>=3.0.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]
